Basic script for testing
TODO: rename the repo to something more helpful

The eventual purpose of this application will be to handle the gathering of userdata, and then storing in the database.

Right now it just serves to make sure the service key functions. The service key just bypasses row level security so it can access all of the data. The idea is that this part of the application will not have any exposed endpoints. Which i hope will help with some semblance of security.